feat: complete SDE-2 portfolio — concurrency proof, performance metrics, architecture diagrams, LLD, README
Core features built:
- Pessimistic locking via DB UniqueConstraint(event_id, seat_id) — 50 concurrent users →
  exactly 1 success + 49 HTTP 409s, zero double-bookings
- Redis caching layer with 16.6× speedup (8.9ms vs 148ms), 62.5% hit rate
- GET /api/v1/metrics endpoint exposing live cache/DB telemetry
- Celery + RabbitMQ async notification pipeline (acks_late=True)
- SQLAlchemy connection pool (size=20, overflow=10) for high concurrency

Proof artefacts added:
- tests/test_concurrency.py  — asyncio.gather 50-user race test (~16s)
- proof/analysis.ipynb       — 4-panel performance dashboard + concurrency bar chart
- proof/architecture.ipynb   — 9 rendered diagrams (HLD, swimlane, LLD class diagram,
                               ERD, tech stack, design decisions, concurrency mechanism,
                               scaling roadmap)
- proof/performance_dashboard.png
- proof/concurrency_proof.png
- proof/lld_class_diagram.png  (UML: 6 entities, 3 enums, BookingService, constraints)
- PROOF.md                   — written proof document with numbers

README rewritten:
- Leads with measurable numbers (16.6×, 62.5%, 37 RPS, <1% error rate)
- Architecture ASCII diagram, service port table, concurrency deep-dive
- Full proof/diagrams reference table, tech stack, scaling roadmap

+++ app/schemas/booking.py
@@ -1,12 +1,13 @@
 from pydantic import BaseModel
 from typing import List
 from datetime import datetime
+from decimal import Decimal
 from .event import Seat  # Re-use the Seat schema we already created
 
 
 class Ticket(BaseModel):
     id: int
-    price: float
+    price: Decimal  # matches Numeric(10,2) — Pydantic serializes Decimal correctly as number
     seat: Seat
 
     class Config:
